refactor: update import paths for PageContext in Dashboard components
