+++ pages/_app.page.tsx
@@ -3,6 +3,7 @@ import { appWithTranslation } from "next-i18next";
 import type { AppProps } from "next/app";
 
 import { SidebarProvider } from "@/context/SidebarContext";
+import { PageProvider } from "@/context/PageContext";
 
 if (process.env.NEXT_PUBLIC_API_MOCKING === "true") {
   import("../mocks").then(({ setupMocks }) => {
@@ -12,9 +13,11 @@ if (process.env.NEXT_PUBLIC_API_MOCKING === "true") {
 
 function App({ Component, pageProps }: AppProps) {
   return (
-    <SidebarProvider>
-      <Component {...pageProps} />
-    </SidebarProvider>
+    <PageProvider>
+      <SidebarProvider>
+        <Component {...pageProps} />
+      </SidebarProvider>
+    </PageProvider>
   );
 }
